pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "wildlife-grad-dashboard"
version = "2.0.0"
description = "Wildlife Graduate Position Intelligence Platform"
authors = [
    {name = "Chris Chizinski", email = "cchizinski2@unl.edu"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "selenium>=4.15.0",
    "beautifulsoup4>=4.12.0",
    "pandas>=2.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "fake-useragent>=1.4.0",
    "webdriver-manager>=4.0.0",
    "scikit-learn>=1.3.0",
    "numpy>=1.24.0",
    "supabase>=2.0.0",
    "aiohttp>=3.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-mock>=3.12.0",
    "pre-commit>=3.6.0",
]

# Tool configurations
[tool.ruff]
target-version = "py39"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "E203",  # whitespace before ':'
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"tests/**/*" = ["S101"]

[tool.mypy]
python_version = "3.9"
warn_return_any = false
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = false
ignore_missing_imports = true

# Exclude data and test files from strict checking
[[tool.mypy.overrides]]
module = [
    "tests.*",
    "data.*",
    "scripts.*",
]
ignore_errors = true

[tool.bandit]
exclude_dirs = ["tests", "data", ".ai-assistant"]
skips = ["B101", "B602", "B603", "B301"]  # Skip common data science issues

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
